Add anthropic web search and api
